meson: Add implicit_include_directories parameter
This fixes build failures with Xcode 12.5, which was released on April 26,
2021. The Xcode release uses a C++ standard library header file named
"version", for the standard library implementation. The standard library
header file conflicts with the file named "VERSION" at the root of the
source directory and makes the build of the "common" library fail.

By explicitly specifying implicit_include_directories to false for
the library, meson will no longer add the root of the source directory
to the search path of angle bracket inclusions (#include <version>).

The header files in the directory is still referencible because it is
specified for quote inclusions (#include "qemu-options.h"). Such a
conflict avoidance was the intention of the commit introducing quote
inclusion path specifications (commit 9edc19c9
"make: switch from -I to -iquote"). The commit also suggests that it
would also help to "catch any future uses of #include <> for internal
headers."

Signed-off-by: Akihiko Odaki <akihiko.odaki@gmail.com>